Moved the fractals structs into a different module
