+++ shared/src/structs/fractal_descriptor.rs
@@ -1,4 +1,7 @@
-use super::prelude::*;
+use super::prelude::{
+    iterated_sin_z::IteratedSinZ, julia::Julia, mandelbrot::Mandelbrot,
+    newton_raphson_z3::NewtonRaphsonZ3, newton_raphson_z4::NewtonRaphsonZ4, *,
+};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub enum FractalDescriptor {
@@ -7,28 +10,4 @@ pub enum FractalDescriptor {
     IteratedSinZ(IteratedSinZ),
     NewtonRaphsonZ3(NewtonRaphsonZ3),
     NewtonRaphsonZ4(NewtonRaphsonZ4),
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Julia {
-    pub c: Complex,
-    pub divergence_threshold_square: f64,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Mandelbrot {}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct IteratedSinZ {
-    pub c: Complex,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct NewtonRaphsonZ3 {
-    pub c: Complex,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct NewtonRaphsonZ4 {
-    pub c: Complex,
 }
