+++ shared/src/structs/fractals/iterated_sin_z.rs
@@ -1,0 +1,6 @@
+use super::super::prelude::*;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct IteratedSinZ {
+    pub c: Complex,
+}
