+++ shared/src/structs/fractals/julia.rs
@@ -1,0 +1,8 @@
+use super::super::prelude::*;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Julia {
+    pub c: Complex,
+    pub divergence_threshold_square: f64,
+}
+
