+++ shared/src/structs/fractals/mandelbrot.rs
@@ -1,0 +1,4 @@
+use super::super::prelude::*;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Mandelbrot {}
