+++ shared/src/structs/fractals/mod.rs
@@ -1,0 +1,5 @@
+pub mod iterated_sin_z;
+pub mod julia;
+pub mod mandelbrot;
+pub mod newton_raphson_z3;
+pub mod newton_raphson_z4;
