+++ shared/src/structs/fractals/newton_raphson_z3.rs
@@ -1,0 +1,7 @@
+use super::super::prelude::*;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct NewtonRaphsonZ3 {
+    pub c: Complex,
+}
+
