+++ shared/src/structs/fractals/newton_raphson_z4.rs
@@ -1,0 +1,6 @@
+use super::super::prelude::*;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct NewtonRaphsonZ4 {
+    pub c: Complex,
+}
