+++ shared/src/structs/mod.rs
@@ -1,5 +1,6 @@
 pub mod complex;
 pub mod fractal_descriptor;
+pub mod fractals;
 pub mod fragment;
 pub mod fragment_request;
 pub mod fragment_result;
@@ -14,6 +15,7 @@ pub mod u8data;
 pub mod prelude {
     pub use super::complex::Complex;
     pub use super::fractal_descriptor::FractalDescriptor;
+    pub use super::fractals::*;
     pub use super::fragment::Fragment;
     pub use super::fragment_request::FragmentRequest;
     pub use super::fragment_result::FragmentResult;
